Add files via upload
This program processes a file with student records and computes the average values for the student's GPA, class standing and age, and writes those values to the output file. As well as computes the minimum, maximum, and standard deviation of the GPA's and writes those values to the output file as well. This program was coded in Microsoft Visual Studio.